update bump script
