+++ package.json
@@ -23,7 +23,7 @@
     "test:all": "pnpm run clean:test && cross-env START_ADAPTER=solid-start-node npm run test",
     "test": "pnpm run clean:test && pnpm --filter solid-start-tests test --",
     "show:test-report": "pnpm --filter solid-start-tests show:test-report",
-    "bump": "node scripts/bump.cjs"
+    "bump": "node scripts/bump.js"
   },
   "devDependencies": {
     "@cloudflare/kv-asset-handler": "^0.2.0",
